added showAllUsers, insertUser SQL requests

+++ src/com/company/dbMethods/interfaces/IMethods.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface IMethods {
     boolean createTableUsers();
     boolean createTableSubjects();
-    boolean insertUser(String name, int gpa);
+    boolean insertUser(String name, double gpa);
     boolean insertSubject(int id, String subject_name, int credits, int grade);
     int getId(String username);
     List<User> getAllUsers();
